New GEOjson to replace old Geojson
Only change is the numerical "id" values for the states is now changed such that the "id" is the string containing the state name, first letter capitalized. This makes graphing on Plotly/Mapbox easier as you don't have to convert or make a new column just for the numerical id values.